chore(page): update scripts
