+++ package.json
@@ -37,11 +37,13 @@
     "@commitlint/config-conventional": "^8.1.0",
     "@commitlint/config-lerna-scopes": "^8.1.0",
     "@types/dotenv": "^6.1.0",
+    "@types/fs-extra": "^8.0.0",
     "@types/graphql": "^14.2.3",
     "@types/jest": "^24.0.0",
     "@types/jsonwebtoken": "^8.3.0",
     "@types/mongodb": "^3.1.18",
     "@types/node-fetch": "^2.1.4",
+    "@types/node-sass": "^4.11.0",
     "autoprefixer": "^9.4.4",
     "babel-jest": "^24.0.0",
     "babel-loader": "^8.0.5",
