+++ packages/page/config/env.js
@@ -22,9 +22,6 @@ var dotenvFiles = [
   paths.dotenv,
 ].filter(Boolean)
 
-// Load environment variables from .env* files.
-// Suppress warnings using silent if this file is missing.
-// dotenv will never modify any environment variables that have already been set.
 // https://github.com/motdotla/dotenv
 // https://github.com/motdotla/dotenv-expand
 dotenvFiles.forEach(dotenvFile => {
@@ -37,13 +34,11 @@ dotenvFiles.forEach(dotenvFile => {
   }
 })
 
-// We support resolving modules according to `NODE_PATH`.
 // This lets you use absolute paths in imports inside large monorepos:
 // https://github.com/facebookincubator/create-react-app/issues/253.
 // It works similar to `NODE_PATH` in Node itself:
 // https://nodejs.org/api/modules.html#modules_loading_from_the_global_folders
 // Note that unlike in Node, only *relative* paths from `NODE_PATH` are honored.
-// Otherwise, we risk importing Node.js core modules into an app instead of Webpack shims.
 // https://github.com/facebookincubator/create-react-app/issues/1023#issuecomment-265344421
 const appDirectory = fs.realpathSync(process.cwd())
 process.env.NODE_PATH = (process.env.NODE_PATH || '')
@@ -65,16 +60,11 @@ function getClientEnvironment (publicUrl) {
         return env
       },
       {
-        // Useful for determining whether we’re running in production mode.
         NODE_ENV: process.env.NODE_ENV || 'development',
-        // Useful for resolving the correct path to static assets in `public`.
         // For example, <img src={process.env.PUBLIC_URL + '/img/logo.png'} />.
-        // This should only be used as an escape hatch. Normally you would put
-        // images into the `src` and `import` them in code to get their paths.
         PUBLIC_URL: publicUrl,
       }
     )
-  // Stringify all values so we can feed into Webpack DefinePlugin
   const stringified = {
     'process.env': Object.keys(raw).reduce(
       (env, key) => {
