+++ packages/page/package.json
@@ -7,7 +7,7 @@
     "start": "NODE_ENV=development npm-run-all -p watch-css start-js",
     "build-js": "node scripts/build.js",
     "build": "npm-run-all build-css build-js",
-    "build-css": "node scripts/build-css.js",
+    "build-css": "ts-node -O '{\"module\": \"commonjs\"}' scripts/build-css.ts",
     "watch-css": "npm run build-css && node-sass-chokidar --include-path ./node_modules ./src/scss -o ./src/css --watch --recursive",
     "test": "node scripts/test.js",
     "ci": "node scripts/test.js --ci --coverage"
@@ -23,6 +23,7 @@
   "devDependencies": {
     "@babel/plugin-transform-react-jsx": "^7.3.0",
     "node-sass-chokidar": "^1.3.5",
-    "npm-run-all": "^4.1.5"
+    "npm-run-all": "^4.1.5",
+    "ts-node": "^8.0.3"
   }
 }
