+++ packages/page/scripts/start.js
@@ -30,7 +30,6 @@ const createDevServerConfig = require('../config/webpackDevServer.config')
 const useYarn = fs.existsSync(paths.yarnLockFile)
 const isInteractive = process.stdout.isTTY
 
-// Warn and crash if required files are missing
 if (!checkRequiredFiles([paths.appHtml, paths.appIndexJs])) {
   process.exit(1)
 }
@@ -45,25 +44,22 @@ choosePort(HOST, DEFAULT_PORT)
     const protocol = process.env.HTTPS === 'true' ? 'https' : 'http'
     const appName = require(paths.appPackageJson).name
     const urls = prepareUrls(protocol, HOST, port)
-    // Create a webpack compiler that is configured with custom messages.
+
     const compiler = createCompiler({webpack, config, appName, urls, useYarn})
-    // Load proxy config
+
     const proxySetting = require(paths.appPackageJson).proxy
     const proxyConfig = prepareProxy(proxySetting, paths.appPublic)
-    // Serve webpack assets generated by the compiler over a web sever.
+
     const serverConfig = createDevServerConfig(
       proxyConfig,
       urls.lanUrlForConfig
     )
     const devServer = new WebpackDevServer(compiler, serverConfig)
-    // Launch WebpackDevServer.
+
     devServer.listen(port, HOST, err => {
-      if (err) {
-        return console.log(err)
-      }
-      if (isInteractive) {
-        clearConsole()
-      }
+      if (err) return console.log(err)
+      if (isInteractive) clearConsole()
+
       console.log(chalk.cyan('Starting the development server...\n'))
       // openBrowser(urls.localUrlForBrowser)
     })
@@ -76,8 +72,6 @@ choosePort(HOST, DEFAULT_PORT)
     })
   })
   .catch(err => {
-    if (err && err.message) {
-      console.log(err.message)
-    }
+    if (err && err.message) console.error(err.message)
     process.exit(1)
   })
